# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "smista-router HTTP API",
    "description": "Local-first deterministic model-routing API. See docs/api/http-api.md.",
    "contact": {
      "name": "Christian Visintin",
      "email": "christian.visintin@veeso.dev"
    },
    "license": {
      "name": "Elastic-2.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Handles `GET /api/v1/auth/me`.",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "The authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/auth/sign-out": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Handles `POST /api/v1/auth/sign-out`.",
        "operationId": "signOut",
        "responses": {
          "200": {
            "description": "Session revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignOutResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Envelope wrapping the error body under an `error` key.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody",
            "description": "The structured error payload."
          }
        }
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "The structured payload of an [`ApiError`].\n\n`details` carries machine-readable context (such as the offending provider\nand model) and is omitted when there is none. It must never contain secrets.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code, for example `missing_provider_credentials`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, actionable description of what went wrong."
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "description": "Optional structured context; never carries secrets."
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "description": "Response to `GET /auth/me`, identifying the authenticated user.\n\nCarries only the caller's user ID; it lists no sessions and exposes no\nsecret values. To enumerate sessions, use `GET /sessions` instead.",
        "required": [
          "user_id"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "description": "The authenticated user's unique identifier."
          }
        }
      },
...
      "SignOutResponse": {
        "type": "object",
        "description": "Response to `POST /auth/sign-out`, confirming token revocation.",
        "required": [
          "revoked"
        ],
        "properties": {
          "revoked": {
            "type": "boolean",
            "description": "Whether the session token was revoked."
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Smista-Api-Key",
        "description": "User API key, presented only at sign-in"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from POST /api/v1/auth/sign-in"
      }
    }
  },
...
}